Refactor packages in cache module
