+++ shardingsphere-features/shardingsphere-sharding/shardingsphere-sharding-plugin/shardingsphere-sharding-cache/src/main/java/org/apache/shardingsphere/sharding/cache/checker/algorithm/CacheableShardingAlgorithmChecker.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sharding.cache.checker;
+package org.apache.shardingsphere.sharding.cache.checker.algorithm;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
