+++ shardingsphere-features/shardingsphere-sharding/shardingsphere-sharding-plugin/shardingsphere-sharding-cache/src/main/java/org/apache/shardingsphere/sharding/cache/route/cache/ShardingRouteCacheValue.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sharding.cache.route;
+package org.apache.shardingsphere.sharding.cache.route.cache;
 
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
